Moving logic under `get_tests` to `nodes`
