+++ main/process/__init__.py
@@ -3,4 +3,4 @@
 
 
 # trunk-ignore(ruff/F401)
-from . import assertions, casts, environment, get_tests, locations, patches
+from . import assertions, casts, environment, locations, nodes, patches
