+++ main/process/nodes/combine_fields.py
@@ -5,7 +5,6 @@
 from types import SimpleNamespace as sns
 from typing import Any
 
-from main import utils
 from main.utils import get_config
